Small adjustments to error messages, and check for selection of attribute from child Types.
