+++ src/Highway.Data.Test/InMemory/BugTests/CreatePollingDeviceSpecification.cs
@@ -1,0 +1,7 @@
+namespace Highway.Data.Tests.InMemory.BugTests
+{
+    public class CreatePollingDeviceSpecification
+    {
+        public string DeviceModel { get; set; }
+    }
+}
