+++ src/Highway.Data.Test/InMemory/BugTests/DeviceModel.cs
@@ -1,0 +1,11 @@
+namespace Highway.Data.Tests.InMemory.BugTests
+{
+    public class DeviceModel : IIdentifiable<long>
+    {
+        public string Code { get; set; }
+
+        public long Id { get; set; }
+
+        public string Name { get; set; }
+    }
+}
