+++ src/Highway.Data.Test/InMemory/BugTests/EntityType.cs
@@ -1,0 +1,9 @@
+namespace Highway.Data.Tests.InMemory.BugTests
+{
+    public class EntityType
+    {
+        public string EntityTypeName { get; set; }
+
+        public int Id { get; set; }
+    }
+}
