+++ src/Highway.Data.Test/InMemory/BugTests/ReadonlyProperties/ParentsWithIgnoredReadonlyChild/Child.cs
@@ -1,9 +1,6 @@
 ﻿namespace Highway.Data.Test.InMemory.BugTests.ReadonlyProperties.ParentsWithIgnoredReadonlyChild
 {
-    public class Child : IIdentifiable<long>
+    public class Child : ChildBase
     {
-        public long Id { get; set; }
-
-        public string Name { get; set; }
     }
 }
