+++ src/Highway.Data.Test/InMemory/BugTests/ReadonlyProperties/ParentsWithReadonlyChild/Child.cs
@@ -1,9 +1,6 @@
 ﻿namespace Highway.Data.Test.InMemory.BugTests.ReadonlyProperties.ParentsWithReadonlyChild
 {
-    public class Child : IIdentifiable<long>
+    public class Child : ChildBase
     {
-        public long Id { get; set; }
-
-        public string Name { get; set; }
     }
 }
