+++ src/Highway.Data.Test/InMemory/BugTests/ReadonlyProperties/ParentsWithReadonlyChildren/Child.cs
@@ -1,9 +1,6 @@
 ﻿namespace Highway.Data.Test.InMemory.BugTests.ReadonlyProperties.ParentsWithReadonlyChildren
 {
-    public class Child : IIdentifiable<long>
+    public class Child : ChildBase
     {
-        public long Id { get; set; }
-
-        public string Name { get; set; }
     }
 }
