+++ src/Highway.Data.Test/InMemory/BugTests/ReadonlyProperties/ParentsWithReadonlyChildren/ChildBase.cs
@@ -1,0 +1,9 @@
+namespace Highway.Data.Test.InMemory.BugTests.ReadonlyProperties.ParentsWithReadonlyChildren
+{
+    public class ChildBase : IIdentifiable<long>
+    {
+        public long Id { get; set; }
+
+        public string Name { get; set; }
+    }
+}
